requirements.txt
Python 3.5.2 for training simulation

h5py (2.7.1)
imageio (2.3.0)
Keras (2.1.3)
matplotlib (1.5.3)
numpy (1.14.0)
Pillow (4.3.0)
scipy (1.0.0)
seaborn (0.8.1)
tensorflow (1.5.0)
tensorflow-gpu (1.4.0)

Python 2.7.12 for NAOqi Python SDK (Version 2.1.2 Linux 64) (robot control)
Package                       Version     
----------------------------- ------------     
h5py                          2.8.0           
imageio                       2.3.0        
Keras                         2.2.2       
Keras-Applications            1.0.4       
Keras-Preprocessing           1.0.2        
matplotlib                    2.2.2       
numpy                         1.14.5       
Pillow                        5.1.0    
scipy                         1.1.0      
tensorflow                    1.5.0    
